!!! TASK: Remove driver for Elasticsearch Version 1 and Version 2 Both Versions are EOL and not longer supported by Elastic. Removing the support for this versions makes testing and improving the package a lot easier
