--- Classes/Driver/Version1/Mapping/NodeTypeMappingBuilder.php
@@ -1,89 +1,0 @@
-<?php
-namespace Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version1\Mapping;
-
-/*
- * This file is part of the Flowpack.ElasticSearch.ContentRepositoryAdaptor package.
- *
- * (c) Contributors of the Neos Project - www.neos.io
- *
- * This package is Open Source Software. For the full copyright and license
- * information, please view the LICENSE file which was distributed with this
- * source code.
- */
-
-use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\AbstractNodeTypeMappingBuilder;
-use Flowpack\ElasticSearch\Domain\Model\Index;
-use Flowpack\ElasticSearch\Domain\Model\Mapping;
-use Flowpack\ElasticSearch\Mapping\MappingCollection;
-use Neos\ContentRepository\Domain\Model\NodeType;
-use Neos\Error\Messages\Result;
-use Neos\Error\Messages\Warning;
-use Neos\Flow\Annotations as Flow;
-
-/**
- * NodeTypeMappingBuilder for Elasticsearch version 1.x
- *
- * @Flow\Scope("singleton")
- */
-class NodeTypeMappingBuilder extends AbstractNodeTypeMappingBuilder
-{
-    /**
-     * Builds a Mapping Collection from the configured node types
-     *
-     * @param Index $index
-     * @return MappingCollection<\Flowpack\ElasticSearch\Domain\Model\Mapping>
-     */
-    public function buildMappingInformation(Index $index)
-    {
-        $this->lastMappingErrors = new Result();
-
-        $mappings = new MappingCollection(MappingCollection::TYPE_ENTITY);
-
-        /** @var NodeType $nodeType */
-        foreach ($this->nodeTypeManager->getNodeTypes() as $nodeTypeName => $nodeType) {
-            if ($nodeTypeName === 'unstructured' || $nodeType->isAbstract()) {
-                continue;
-            }
-
-            $type = $index->findType($this->convertNodeTypeNameToMappingName($nodeTypeName));
-            $mapping = new Mapping($type);
-            $fullConfiguration = $nodeType->getFullConfiguration();
-            if (isset($fullConfiguration['search']['elasticSearchMapping'])) {
-                $mapping->setFullMapping($fullConfiguration['search']['elasticSearchMapping']);
-            }
-
-            // https://www.elastic.co/guide/en/elasticsearch/reference/2.4/dynamic-templates.html
-            // 'not_analyzed' is necessary
-            $mapping->addDynamicTemplate('dimensions', [
-                'path_match' => '__dimensionCombinations.*',
-                'match_mapping_type' => 'string',
-                'mapping' => [
-                    'type' => 'string',
-                    'index' => 'not_analyzed'
-                ]
-            ]);
-            $mapping->setPropertyByPath('__dimensionCombinationHash', [
-                'type' => 'string',
-                'index' => 'not_analyzed'
-            ]);
-
-            foreach ($nodeType->getProperties() as $propertyName => $propertyConfiguration) {
-                if (isset($propertyConfiguration['search']) && isset($propertyConfiguration['search']['elasticSearchMapping'])) {
-                    if (is_array($propertyConfiguration['search']['elasticSearchMapping'])) {
-                        $mapping->setPropertyByPath($propertyName, $propertyConfiguration['search']['elasticSearchMapping']);
-                    }
-                } elseif (isset($propertyConfiguration['type']) && isset($this->defaultConfigurationPerType[$propertyConfiguration['type']]['elasticSearchMapping'])) {
-                    if (is_array($this->defaultConfigurationPerType[$propertyConfiguration['type']]['elasticSearchMapping'])) {
-                        $mapping->setPropertyByPath($propertyName, $this->defaultConfigurationPerType[$propertyConfiguration['type']]['elasticSearchMapping']);
-                    }
-                } else {
-                    $this->lastMappingErrors->addWarning(new Warning('Node Type "' . $nodeTypeName . '" - property "' . $propertyName . '": No ElasticSearch Mapping found.'));
-                }
-            }
-
-            $mappings->add($mapping);
-        }
-
-        return $mappings;
-    }
-}
