--- Classes/Driver/Version2/Query/FunctionScoreQuery.php
@@ -1,18 +1,0 @@
-<?php
-namespace Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version2\Query;
-
-/*
- * This file is part of the Flowpack.ElasticSearch.ContentRepositoryAdaptor package.
- *
- * (c) Contributors of the Neos Project - www.neos.io
- *
- * This package is Open Source Software. For the full copyright and license
- * information, please view the LICENSE file which was distributed with this
- * source code.
- */
-
-use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version1;
-
-class FunctionScoreQuery extends Version1\Query\FunctionScoreQuery
-{
-}
