+++ Classes/Driver/Version5/DocumentDriver.php
@@ -11,14 +11,90 @@ namespace Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version5;
  * source code.
  */
 
-use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version2;
 use Neos\Flow\Annotations as Flow;
+use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\AbstractDriver;
+use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\DocumentDriverInterface;
+use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\NodeTypeMappingBuilderInterface;
+use Flowpack\ElasticSearch\Domain\Model\Index;
+use Neos\ContentRepository\Domain\Model\NodeInterface;
+use Neos\ContentRepository\Domain\Model\NodeType;
 
 /**
  * Document driver for Elasticsearch version 5.x
  *
  * @Flow\Scope("singleton")
  */
-class DocumentDriver extends Version2\DocumentDriver
+class DocumentDriver extends AbstractDriver implements DocumentDriverInterface
 {
+    /**
+     * @Flow\Inject
+     * @var NodeTypeMappingBuilderInterface
+     */
+    protected $nodeTypeMappingBuilder;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function delete(NodeInterface $node, $identifier)
+    {
+        return [
+            [
+                'delete' => [
+                    '_type' => $this->nodeTypeMappingBuilder->convertNodeTypeNameToMappingName($node->getNodeType()),
+                    '_id' => $identifier
+                ]
+            ]
+        ];
+    }
+
+    /**
+     * {@inheritdoc}
+     * @throws \Flowpack\ElasticSearch\Exception
+     */
+    public function deleteDuplicateDocumentNotMatchingType(Index $index, $documentIdentifier, NodeType $nodeType)
+    {
+        $result = $index->request('GET', '/_search?scroll=1m', [], json_encode([
+            'sort' => ['_doc'],
+            'query' => [
+                'bool' => [
+                    'must' => [
+                        'ids' => [
+                            'values' => [$documentIdentifier]
+                        ]
+                    ],
+                    'must_not' => [
+                        'term' => [
+                            '_type' => $this->nodeTypeMappingBuilder->convertNodeTypeNameToMappingName($nodeType->getName())
+                        ]
+                    ]
+                ]
+            ]
+        ]));
+        $treatedContent = $result->getTreatedContent();
+        $scrollId = $treatedContent['_scroll_id'];
+        $mapHitToDeleteRequest = function ($hit) {
+            return json_encode([
+                'delete' => [
+                    '_type' => $hit['_type'],
+                    '_id' => $hit['_id']
+                ]
+            ]);
+        };
+        $bulkRequest = [];
+        while (isset($treatedContent['hits']['hits']) && $treatedContent['hits']['hits'] !== []) {
+            $hits = $treatedContent['hits']['hits'];
+            $bulkRequest = array_merge($bulkRequest, array_map($mapHitToDeleteRequest, $hits));
+            $result = $index->request('GET', '/_search/scroll?scroll=1m', [], $scrollId, false);
+            $treatedContent = $result->getTreatedContent();
+        }
+        $this->logger->log(sprintf('NodeIndexer: Check duplicate nodes for %s (%s), found %d document(s)', $documentIdentifier, $nodeType->getName(), count($bulkRequest)), LOG_DEBUG, null, 'ElasticSearch (CR)');
+        if ($bulkRequest !== []) {
+            $index->request('POST', '/_bulk', [], implode("\n", $bulkRequest) . "\n");
+        }
+        $this->searchClient->request('DELETE', '/_search/scroll', [], json_encode([
+            'scroll_id' => [
+                $scrollId
+            ]
+        ]));
+    }
 }
