+++ Classes/Driver/Version5/IndexDriver.php
@@ -11,14 +11,55 @@ namespace Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version5;
  * source code.
  */
 
-use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version1;
 use Neos\Flow\Annotations as Flow;
+use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\AbstractDriver;
+use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\IndexDriverInterface;
+use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Exception;
 
 /**
  * Index management driver for Elasticsearch version 5.x
  *
  * @Flow\Scope("singleton")
  */
-class IndexDriver extends Version1\IndexDriver
+class IndexDriver extends AbstractDriver implements IndexDriverInterface
 {
+    /**
+     * {@inheritdoc}
+     */
+    public function aliasActions(array $actions)
+    {
+        $this->searchClient->request('POST', '/_aliases', [], \json_encode(['actions' => $actions]));
+    }
+
+    /**
+     * {@inheritdoc}
+     * @throws Exception
+     */
+    public function deleteIndex(string $index)
+    {
+        $response = $this->searchClient->request('HEAD', '/' . $index);
+        if ($response->getStatusCode() === 200) {
+            $response = $this->searchClient->request('DELETE', '/' . $index);
+            if ($response->getStatusCode() !== 200) {
+                throw new Exception('The index "' . $index . '" could not be deleted. (return code: ' . $response->getStatusCode() . ')', 1395419177);
+            }
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     * @throws Exception
+     */
+    public function indexesByAlias(string $alias)
+    {
+        $response = $this->searchClient->request('GET', '/_alias/' . $alias);
+        if ($response->getStatusCode() !== 200 && $response->getStatusCode() !== 404) {
+            throw new Exception('The alias "' . $alias . '" was not found with some unexpected error... (return code: ' . $response->getStatusCode() . ')', 1383650137);
+        }
+
+        // return empty array if content from response cannot be read as an array
+        $treatedContent = $response->getTreatedContent();
+
+        return is_array($treatedContent) ? array_keys($treatedContent) : [];
+    }
 }
