+++ Classes/Driver/Version5/IndexerDriver.php
@@ -11,18 +11,20 @@ namespace Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version5;
  * source code.
  */
 
-use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version2;
 use Flowpack\ElasticSearch\Domain\Model\Document as ElasticSearchDocument;
 use Neos\ContentRepository\Domain\Model\NodeInterface;
 use Neos\Flow\Annotations as Flow;
+use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\AbstractIndexerDriver;
+use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\IndexerDriverInterface;
 
 /**
  * Indexer driver for Elasticsearch version 5.x
  *
  * @Flow\Scope("singleton")
  */
-class IndexerDriver extends Version2\IndexerDriver
+class IndexerDriver extends AbstractIndexerDriver implements IndexerDriverInterface
 {
+
     /**
      * {@inheritdoc}
      */
