+++ Configuration/Settings.yaml
@@ -2,82 +2,8 @@ Flowpack:
   ElasticSearch:
     ContentRepositoryAdaptor:
       driver:
-        version: '1.x'
+        version: '5.x'
         mapping:
-          '1.x':
-            query:
-              className: 'Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version1\Query\FilteredQuery'
-              arguments:
-                request: &request_version1
-                  query:
-                    filtered:
-                      query:
-                        bool:
-                          must:
-                            -
-                              match_all: {  }
-                      filter:
-                        bool:
-                          must: {  }
-                          should: {  }
-                          must_not:
-                            -
-                              term:
-                                _hidden: true
-                            -
-                              range:
-                                _hiddenBeforeDateTime:
-                                  gt: now
-                            -
-                              range:
-                                _hiddenAfterDateTime:
-                                  lt: now
-                  fields:
-                    - __path
-                unsupportedFieldsInCountRequest: &unsupportedFieldsInCountRequest_version1
-                  - 'fields'
-                  - 'sort'
-                  - 'from'
-                  - 'size'
-                  - 'highlight'
-                  - 'aggs'
-                  - 'aggregations'
-                  - 'suggest'
-
-            document:
-              className: 'Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version1\DocumentDriver'
-            indexer:
-              className: 'Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version1\IndexerDriver'
-            indexManagement:
-              className: 'Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version1\IndexDriver'
-            request:
-              className: 'Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version1\RequestDriver'
-            system:
-              className: Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version1\SystemDriver
-            nodeTypeMappingBuilder:
-              className: 'Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version1\Mapping\NodeTypeMappingBuilder'
-
-          '2.x':
-            query:
-              className: 'Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version2\Query\FilteredQuery'
-              arguments:
-                request:
-                  <<: *request_version1
-                unsupportedFieldsInCountRequest:
-                  <<: *unsupportedFieldsInCountRequest_version1
-            document:
-              className: 'Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version2\DocumentDriver'
-            indexer:
-              className: 'Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version2\IndexerDriver'
-            indexManagement:
-              className: 'Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version2\IndexDriver'
-            request:
-              className: 'Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version2\RequestDriver'
-            system:
-              className: Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version2\SystemDriver
-            nodeTypeMappingBuilder:
-              className: Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version2\Mapping\NodeTypeMappingBuilder
-
           5.x:
             query:
               className: Flowpack\ElasticSearch\ContentRepositoryAdaptor\Driver\Version5\Query\FilteredQuery
