--- Configuration/Testing/ElasticVersion1/Settings.yaml
@@ -1,5 +1,0 @@
-Flowpack:
-  ElasticSearch:
-    ContentRepositoryAdaptor:
-      driver:
-        version: '1.x'
