--- Configuration/Testing/ElasticVersion2/Settings.yaml
@@ -1,5 +1,0 @@
-Flowpack:
-  ElasticSearch:
-    ContentRepositoryAdaptor:
-      driver:
-        version: '2.x'
